requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
langchain_text_splitters
langchain-openai
chromadb
faiss-cpu
beautifulsoup4
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
langchain-chroma
langchain_huggingface
sentence-transformers
bs4
